// frontend/src/api/models/CourseResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CourseResponse
 */
export interface CourseResponse {
    /**
     * 
     * @type {string}
     * @memberof CourseResponse
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof CourseResponse
     */
    description?: string | null;
    /**
     * 
     * @type {number}
     * @memberof CourseResponse
     */
    categoryId: number;
    /**
     * 
     * @type {number}
     * @memberof CourseResponse
     */
    teacherId: number;
    /**
     * 
     * @type {number}
     * @memberof CourseResponse
     */
    deadlineInDays: number | null;
    /**
     * 
     * @type {boolean}
     * @memberof CourseResponse
     */
    isActive?: boolean;
    /**
     * 
     * @type {number}
     * @memberof CourseResponse
     */
    courseId: number;
}

/**
 * Check if a given object implements the CourseResponse interface.
 */
export function instanceOfCourseResponse(value: object): value is CourseResponse {
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('categoryId' in value) || value['categoryId'] === undefined) return false;
    if (!('teacherId' in value) || value['teacherId'] === undefined) return false;
    if (!('deadlineInDays' in value) || value['deadlineInDays'] === undefined) return false;
    if (!('courseId' in value) || value['courseId'] === undefined) return false;
    return true;
}

export function CourseResponseFromJSON(json: any): CourseResponse {
    return CourseResponseFromJSONTyped(json, false);
}

export function CourseResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CourseResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'],
        'description': json['description'] == null ? undefined : json['description'],
        'categoryId': json['category_id'],
        'teacherId': json['teacher_id'],
        'deadlineInDays': json['deadline_in_days'],
        'isActive': json['is_active'] == null ? undefined : json['is_active'],
        'courseId': json['course_id'],
    };
}

export function CourseResponseToJSON(json: any): CourseResponse {
    return CourseResponseToJSONTyped(json, false);
}

export function CourseResponseToJSONTyped(value?: CourseResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'description': value['description'],
        'category_id': value['categoryId'],
        'teacher_id': value['teacherId'],
        'deadline_in_days': value['deadlineInDays'],
        'is_active': value['isActive'],
        'course_id': value['courseId'],
    };
}